Change styles in the accordion and pagination by defaul of bootstrap
