+++ src/components/pagination/index.js
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+import Pagination from "./pagination";
+export default styled(Pagination)`
+  .pagination {
+    --bs-pagination-focus-color: #0b7285;
+    --bs-pagination-color: #fff;
+    --bs-pagination-bg: #495057;
+    --bs-pagination-border-color: #0b7285;
+    --bs-pagination-hover-bg: #343a40;
+    --bs-pagination-hover-border-color: #0b7285;
+    --bs-pagination-hover-color: #e3fafc;
+
+    --bs-pagination-disabled-color: #343a40;
+    --bs-pagination-disabled-bg: #495057;
+    --bs-pagination-disabled-border-color: #343a40;
+  }
+`;
